requirements.txt
django==4.1.3
django-bootstrap5==22.2
django-bootstrap-datepicker-plus==5.0.0